Add new pep440-git-full mode
This mode is a further specialization of the pep440-git mode. In this
mode, if the git describe output includes the commit ID (SHA-1), it will
be included in the local portion of the generated version. Additionally,
if git describe was run with the --dirty option, it will also include
that marker in the local version.

By embedding this information in the local version portion, two goals
are achieved:

1. Ambiguity between two commits that are the same distance from the tag
   is no longer present.
2. Local uncommitted changes are considered newer than the last
   committed changes.

Examples (with post_mode of ".post"):
    1.2-1-bcde234 becomes 1.2.post1+gbcde234
    1.2-1-bcde234-dirty becomes 1.2.post1+gbcde234.dirty